requirements.txt
# C2 Server & Builder
customtkinter
Flask
requests
pyinstaller

# Payload
psutil
pycryptodomex
pywin32
requests